drone_vision/human_generator/requirements.txt
# PyTorch with CUDA 12.1 (for RTX 2080 Ti)
torch==2.3.0+cu121
torchvision==0.18.0+cu121
torchaudio==2.3.0+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

# Hugging Face + LoRA stack
diffusers==0.24.0
huggingface_hub==0.19.4
transformers==4.33.3
accelerate==0.21.0
peft==0.5.0
datasets==2.12.0

# Vision, data, and training tools
opencv-python
albumentations
imageio
scikit-learn
matplotlib
Pillow>=9.0.0
numpy>=1.23.0
scipy>=1.10.0

# Utilities and logging
tqdm>=4.42.1
pyyaml>=6.0
requests>=2.28
fsspec>=2023.5.0
filelock>=3.12
packaging>=25.0
typing-extensions>=4.0
tensorboard
wandb
